Refactor feature_structure docs and lib/core architecture
- **Feature Structure**:
  - Standardized documentation to Clean Architecture (Data, Domain, Presentation, Application).
  - Detailed explanation of each layer in `README.md` files.
  - Removed `Infrastructure` from feature template (it belongs to Core).
- **Core Refactor**:
  - Moved DTOs (`BaseResponse`, `BaseMessage`) from Domain to `lib/core/data/models`.
  - Moved App State (`RestState`, `UploadState`) from Domain to `lib/core/application/state`.
  - Moved Network Config (`HttpOverrides`) to `lib/core/infrastructure/network/overrides`.
  - Updated imports in `bootstrap.dart`.
