+++ lib/core/application/state/rest_state.dart
