+++ lib/core/application/state/upload_state.dart
