+++ lib/core/data/models/base_message.dart
